improved logic to avoid duplicate instances
