+++ src/main/java/org/folio/dew/batch/bulkedit/jobs/processidentifiers/BulkEditInstanceIdentifiersJobConfig.java
@@ -5,7 +5,7 @@ import lombok.RequiredArgsConstructor;
 import org.folio.dew.batch.CsvListFileWriter;
 import org.folio.dew.batch.JsonListFileWriter;
 import org.folio.dew.batch.JobCompletionNotificationListener;
-import org.folio.dew.batch.bulkedit.jobs.BulkEditInstanceListProcessor;
+import org.folio.dew.batch.bulkedit.jobs.BulkEditInstanceProcessor;
 import org.folio.dew.domain.dto.ExportType;
 import org.folio.dew.domain.dto.ItemIdentifier;
 import org.folio.dew.domain.dto.InstanceFormat;
@@ -19,7 +19,6 @@ import org.springframework.batch.core.launch.support.RunIdIncrementer;
 import org.springframework.batch.core.repository.JobRepository;
 import org.springframework.batch.core.step.builder.StepBuilder;
 import org.springframework.batch.item.file.FlatFileItemReader;
-import org.springframework.batch.item.support.CompositeItemProcessor;
 import org.springframework.batch.item.support.CompositeItemWriter;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
@@ -38,14 +37,12 @@ import static org.folio.dew.utils.Constants.JOB_NAME_POSTFIX_SEPARATOR;
 @Configuration
 @RequiredArgsConstructor
 public class BulkEditInstanceIdentifiersJobConfig {
-
-  private final BulkEditInstanceListProcessor bulkEditInstanceListProcessor;
-  private final InstanceFetcher instanceFetcher;
+  private final BulkEditInstanceProcessor bulkEditInstanceProcessor;
   private final BulkEditSkipListener bulkEditSkipListener;
 
   @Bean
   public Job bulkEditProcessInstanceIdentifiersJob(JobCompletionNotificationListener listener, Step bulkEditInstanceStep,
-                                                   JobRepository jobRepository) {
+    JobRepository jobRepository) {
     return new JobBuilder(ExportType.BULK_EDIT_IDENTIFIERS + JOB_NAME_POSTFIX_SEPARATOR + INSTANCE.getValue(), jobRepository)
       .incrementer(new RunIdIncrementer())
       .listener(listener)
@@ -56,13 +53,13 @@ public class BulkEditInstanceIdentifiersJobConfig {
 
   @Bean
   public Step bulkEditInstanceStep(FlatFileItemReader<ItemIdentifier> csvItemIdentifierReader,
-                               CompositeItemWriter<List<InstanceFormat>> compositeInstanceListWriter,
-                               ListIdentifiersWriteListener<InstanceFormat> listIdentifiersWriteListener, JobRepository jobRepository,
-                               PlatformTransactionManager transactionManager) {
+    CompositeItemWriter<List<InstanceFormat>> compositeInstanceListWriter,
+    ListIdentifiersWriteListener<InstanceFormat> listIdentifiersWriteListener, JobRepository jobRepository,
+    PlatformTransactionManager transactionManager) {
     return new StepBuilder("bulkEditInstanceStep", jobRepository)
       .<ItemIdentifier, List<InstanceFormat>> chunk(CHUNKS, transactionManager)
       .reader(csvItemIdentifierReader)
-      .processor(identifierInstanceProcessor())
+      .processor(bulkEditInstanceProcessor)
       .faultTolerant()
       .skipLimit(1_000_000)
       .processorNonTransactional() // Required to avoid repeating BulkEditItemProcessor#process after skip.
@@ -71,13 +68,6 @@ public class BulkEditInstanceIdentifiersJobConfig {
       .writer(compositeInstanceListWriter)
       .listener(listIdentifiersWriteListener)
       .build();
-  }
-
-  @Bean
-  public CompositeItemProcessor<ItemIdentifier, List<InstanceFormat>> identifierInstanceProcessor() {
-    var processor = new CompositeItemProcessor<ItemIdentifier, List<InstanceFormat>>();
-    processor.setDelegates(Arrays.asList(instanceFetcher, bulkEditInstanceListProcessor));
-    return processor;
   }
 
   @Bean
