+++ src/test/resources/mappings/instances-query.json
@@ -68,7 +68,20 @@
     {
       "request": {
         "method": "GET",
-        "url": "/inventory/instances?query=%28identifiers%3D/%40identifierTypeId%3D913300b2-03ed-469a-8179-c1092c991227%20%2212345%22%29"
+        "url": "/inventory/instances?query=id%3D%3D1572796a-b88b-4991-a9f7-2e368217c487&limit=1"
+      },
+      "response": {
+        "status": 200,
+        "body": "{\n    \"instances\": [\n        {\n            \"id\": \"00f10ab9-d845-4334-92d2-ff55862bf4f9\",\n            \"_version\": \"3\",\n            \"hrid\": \"inst000000000022\",\n            \"source\": \"FOLIO\",\n            \"title\": \"American Bar Association journal.\",\n            \"administrativeNotes\": [],\n            \"indexTitle\": \"American Bar Association journal.\",\n            \"parentInstances\": [],\n            \"childInstances\": [],\n            \"isBoundWith\": false,\n            \"alternativeTitles\": [],\n            \"editions\": [],\n            \"series\": [],\n            \"identifiers\": [],\n            \"contributors\": [\n                {\n                    \"authorityId\": null,\n                    \"contributorNameTypeId\": \"d376e36c-b759-4fed-8502-7130d1eeff39\",\n                    \"name\": \"American Bar Association\",\n                    \"contributorTypeId\": \"6e09d47d-95e2-4d8a-831b-f777b8ef6d81\",\n                    \"contributorTypeText\": \"\",\n                    \"primary\": null\n                },\n                {\n                    \"authorityId\": null,\n                    \"contributorNameTypeId\": \"d376e36c-b759-4fed-8502-7130d1eeff39\",\n                    \"name\": \"American Bar Association. Journal\",\n                    \"contributorTypeId\": \"06b2cbd8-66bf-4956-9d90-97c9776365a4\",\n                    \"contributorTypeText\": \"\",\n                    \"primary\": null\n                }\n            ],\n            \"subjects\": [],\n            \"classifications\": [],\n            \"publication\": [],\n            \"publicationFrequency\": [\n                \"Monthly, 1921-83\",\n                \"Quarterly, 1915-20\"\n            ],\n            \"publicationRange\": [\n                \"Began with vol. 1, no. 1 (Jan. 1915); ceased with v. 69, [no.12] (Dec. 1983)\"\n            ],\n            \"electronicAccess\": [],\n            \"instanceTypeId\": \"30fffe0e-e985-4144-b2e2-1e8179bdb41f\",\n            \"instanceFormatIds\": [\n                \"5cb91d15-96b1-4b8a-bf60-ec310538da66\"\n            ],\n            \"physicalDescriptions\": [\n                \"69 v. : ill. ; 23-30 cm.\"\n            ],\n            \"languages\": [\n                \"eng\"\n            ],\n            \"notes\": [],\n            \"previouslyHeld\": false,\n            \"discoverySuppress\": false,\n            \"statisticalCodeIds\": [],\n            \"metadata\": {\n                \"createdDate\": \"2023-11-10T11:54:16.187+00:00\",\n                \"createdByUserId\": \"cffb2565-07fc-470b-86c6-17d8ce14432e\",\n                \"updatedDate\": \"2024-01-05T09:52:26.651+00:00\",\n                \"updatedByUserId\": \"ca6022de-2644-46fe-b6f2-78df15483721\"\n            },\n            \"tags\": {\n                \"tagList\": []\n            },\n            \"natureOfContentTermIds\": [\n\t\t\t\t\"921e6d93-bafb-4a02-b62f-dcd027c45406\"\n\t\t\t],\n            \"precedingTitles\": [],\n            \"succeedingTitles\": []\n        }\n    ],\n    \"totalRecords\": 1\n}",
+        "headers": {
+          "Content-Type": "application/json"
+        }
+      }
+    },
+    {
+      "request": {
+        "method": "GET",
+        "url": "/inventory/instances?query=%28identifiers%3D/%40identifierTypeId%3D913300b2-03ed-469a-8179-c1092c991227%20%2212345%22%29&limit=2147483647"
       },
       "response": {
         "status": 200,
@@ -81,7 +94,7 @@
     {
       "request": {
         "method": "GET",
-        "url": "/inventory/instances?query=%28identifiers%3D/%40identifierTypeId%3D8261054f-be78-422d-bd51-4ed9f33c3422%20%2212345%22%29"
+        "url": "/inventory/instances?query=%28identifiers%3D/%40identifierTypeId%3D8261054f-be78-422d-bd51-4ed9f33c3422%20%2212345%22%29&limit=2147483647"
       },
       "response": {
         "status": 200,
