+++ src/test/resources/output/bulk_edit_instance_by_isbn_json_output.json
@@ -1,1 +1,1 @@
-{"id":"7fbd5d84-62d1-44c6-9c45-6cb173998bbd","_version":31,"hrid":"inst000000000006","matchKey":null,"source":"FOLIO","title":"Bridget Jones's Baby: the diaries","indexTitle":null,"alternativeTitles":[],"editions":["First American Edition"],"series":[],"identifiers":[{"value":"12345","identifierTypeId":"913300b2-03ed-469a-8179-c1092c991227","identifierTypeObject":null},{"value":"12345","identifierTypeId":"8261054f-be78-422d-bd51-4ed9f33c3422","identifierTypeObject":null}],"contributors":[{"name":"Fielding, Helen","contributorTypeId":null,"contributorTypeText":null,"contributorNameTypeId":"2b94c631-fca9-4892-a730-03ee529ffe2a","authorityId":null,"contributorNameType":null,"primary":null}],"subjects":[],"classifications":[],"publication":[],"publicationFrequency":["A frequency description"],"publicationRange":["A publication range"],"publicationPeriod":null,"electronicAccess":[],"instanceTypeId":"6312d172-f0cf-40f6-b27d-9fa8feaf332f","instanceFormatIds":[],"instanceFormats":null,"physicalDescriptions":["219 pages ; 20 cm."],"languages":["eng"],"notes":[],"administrativeNotes":["Cataloging data"],"modeOfIssuanceId":null,"catalogedDate":"2024-01-26","previouslyHeld":true,"staffSuppress":false,"discoverySuppress":false,"statisticalCodeIds":[],"sourceRecordFormat":null,"statusId":null,"statusUpdatedDate":null,"tags":{"tagList":[]},"metadata":{"createdDate":1699617257074,"createdByUserId":"cffb2565-07fc-470b-86c6-17d8ce14432e","createdByUsername":null,"updatedDate":1704810384610,"updatedByUserId":"ca6022de-2644-46fe-b6f2-78df15483721","updatedByUsername":null},"holdingsRecords2":null,"natureOfContentTermIds":[],"issn":null,"isbn":"12345"}
+{"id":"7fbd5d84-62d1-44c6-9c45-6cb173998bbd","_version":31,"hrid":"inst000000000006","matchKey":null,"source":"FOLIO","title":"Bridget Jones's Baby: the diaries","indexTitle":null,"alternativeTitles":[],"editions":["First American Edition"],"series":[],"identifiers":[{"value":"12345","identifierTypeId":"913300b2-03ed-469a-8179-c1092c991227","identifierTypeObject":null},{"value":"12345","identifierTypeId":"8261054f-be78-422d-bd51-4ed9f33c3422","identifierTypeObject":null}],"contributors":[{"name":"Fielding, Helen","contributorTypeId":null,"contributorTypeText":null,"contributorNameTypeId":"2b94c631-fca9-4892-a730-03ee529ffe2a","authorityId":null,"contributorNameType":null,"primary":null}],"subjects":[],"classifications":[],"publication":[],"publicationFrequency":["A frequency description"],"publicationRange":["A publication range"],"publicationPeriod":null,"electronicAccess":[],"instanceTypeId":"6312d172-f0cf-40f6-b27d-9fa8feaf332f","instanceFormatIds":[],"instanceFormats":null,"physicalDescriptions":["219 pages ; 20 cm."],"languages":["eng"],"notes":[],"administrativeNotes":["Cataloging data"],"modeOfIssuanceId":null,"catalogedDate":"2024-01-26","previouslyHeld":true,"staffSuppress":false,"discoverySuppress":false,"statisticalCodeIds":[],"sourceRecordFormat":null,"statusId":null,"statusUpdatedDate":null,"tags":{"tagList":[]},"metadata":{"createdDate":1699617257074,"createdByUserId":"cffb2565-07fc-470b-86c6-17d8ce14432e","createdByUsername":null,"updatedDate":1704810384610,"updatedByUserId":"ca6022de-2644-46fe-b6f2-78df15483721","updatedByUsername":null},"holdingsRecords2":null,"natureOfContentTermIds":[],"ISSN":null,"ISBN":"12345"}
